add client and read manager

+++ src/main/java/server/GenerationId.java
@@ -5,7 +5,7 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 
-import static java.lang.Integer.parseInt;
+import static java.lang.Long.parseLong;
 
 /**
  * A class that creates unique id for each dragon
@@ -18,10 +18,10 @@ public class GenerationId {
      * Generate unique id greater than zero
      * @return unique id (int)
      */
-    public static int generatorId() {
-        int id;
+    public static long generatorId() {
+        long id;
         String generateUUIDNo = String.format("%010d", new BigInteger(UUID.randomUUID().toString().replace("-", ""), 16));
-        id = parseInt(generateUUIDNo.substring(generateUUIDNo.length() - 10), 10);
+        id = parseLong(generateUUIDNo.substring(generateUUIDNo.length() - 10), 10);
         return id;
     }
